Panic when subprocess crashes
Rather than using an exit status of 1.

This is still imperfect, but other errors in running the subprocess
are currently being handled with `expect`, so I think it makes
sense to handle this that way too.

An eventual improvement would be to do the same things some shells
do to make it so the effect of `program` and `exec program`, as
seen in the shell's own parent, are the same: crash ourselves with
the same signal the subprocess crashed with. This will make sense
to do if and why other changes to make `useenv` more `env`-like are
made.